requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
ghp-import==2.1.0
h11==0.14.0
idna==3.4
jedi==0.18.2
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-material==9.1.11
mkdocs-material-extensions==1.1.1
packaging==23.0
parso==0.8.3
pokebase==1.3.0
pydantic==1.10.7
Pygments==2.15.1
pymdown-extensions==9.11
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2023.5.5
requests==2.28.2
six==1.16.0
SnakeMD==0.12.1
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchdog==3.0.0
